requirements.txt
wakeonlan
paho-mqtt
hass-mqtt-things>=0.5.9
